backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
playwright==1.41.2
beautifulsoup4==4.12.3
openai==1.10.0
python-dotenv==1.0.0
pydantic==2.5.3
requests==2.31.0
httpx==0.26.0
